+++ apps/imlib/src/imboy_str.erl
@@ -3,9 +3,9 @@
 -export([startswith/2]).
 
 
+% imboy_str:startswith("abc", "abcdef").
 startswith([], _) -> true;
 startswith([Ch | Rest1], [Ch | Rest2]) ->
         startswith(Rest1, Rest2);
 startswith(_, _) -> false.
 
-% startswith("abc", "abcdef").